adding include folder structure, adjusting graphs functions in different classes, adding RL base class
