+++ include/rl.h
@@ -1,0 +1,11 @@
+#include <iostream>
+
+
+class rl
+{
+public:
+	rl();
+	~rl();
+private:
+	int bandit();
+};
