+++ include/stdafx.h
@@ -1,0 +1,7 @@
+#include <iostream>
+#include <queue>
+#include <algorithm>
+#include <string.h>
+#include <set>
+#include <climits>
+#include <fstream>
