+++ src/rl.cpp
@@ -1,0 +1,18 @@
+#include "rl.h"
+
+
+rl::rl()
+{
+	return 0;
+}
+
+
+rl::~rl()
+{
+	return 0;
+}
+
+int rl::bandit()
+{
+	return 0;
+}
